Fixed bug of double inclusion of javascript; Sliding the questions in survey; Create* special pages are not includable anymore

+++ extensions/votapedia/Sms.php
@@ -62,6 +62,11 @@ class Sms
         $list = array();
         foreach($new as $sms)
         {
+            if($sms['SenderNumber'] == Sms::$balanceNumber)
+            {
+                Sms::processed($sms['ID']);
+                continue;
+            }
             if(in_array($sms['SenderNumber'], Sms::$blackList))
             {
                 Sms::delete($sms['ID']);
